requirements-base.txt
PyYAML==6.0.3
requests==2.32.3
numpy==1.26.4
pandas==2.2.3
pydantic==2.9.2
